architecture remodeling complete
