+++ app/src/main/java/com/Phase4/pathways/ui/fragments/Fragment_android_developer_details.kt
@@ -1,4 +1,4 @@
-package com.Phase4.pathways.views
+package com.Phase4.pathways.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
