+++ app/src/main/java/com/Phase4/pathways/ui/fragments/Fragment_data_scientist_details.kt
@@ -1,4 +1,4 @@
-package com.Phase4.pathways.views
+package com.Phase4.pathways.ui.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -9,21 +9,21 @@ import android.widget.Button
 import androidx.navigation.Navigation
 import com.Phase4.pathways.R
 
-class fragment_cyber_security_details : Fragment() {
+class fragment_data_scientist_details : Fragment() {
     private var bootcampBtn: Button? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_cyber_security_details, container, false)
+        val view = inflater.inflate(R.layout.fragment_data_scientist_details, container, false)
 
         bootcampBtn = view.findViewById(R.id.BootcampBtn)
 
         bootcampBtn?.setOnClickListener(View.OnClickListener {
             Navigation.findNavController(
                 requireView()
-            ).navigate(R.id.action_fragment_cyber_security_details_to_fragment_bootcamp)
+            ).navigate(R.id.action_fragment_data_scientist_details_to_fragment_bootcamp)
         })
         return view
     }
